+++ packages/fluentui/accessibility/src/behaviors/Toolbar/toolbarItemBehavior.ts
@@ -7,7 +7,6 @@ import { buttonBehavior, ButtonBehaviorProps } from '../Button/buttonBehavior';
  * @specification
  * Adds role='button' if element type is other than 'button'. This allows screen readers to handle the component as a button.
  * Adds attribute 'tabIndex=0' if element type is other than 'button'.
- * Adds attribute 'aria-disabled=true' based on the property 'disabled'. This can be overriden by providing 'aria-disabled' property directly to the component.
  * Adds attribute 'aria-haspopup=menu' to 'root' slot if 'hasMenu' property is set.
  * Adds attribute 'aria-haspopup=dialog' to 'root' slot if 'hasPopup' property is set.
  * Triggers 'performClick' action with 'Enter' or 'Spacebar' on 'root'.