+++ packages/fluentui/docs/src/examples/components/Tooltip/Usage/TooltipExampleDisabledTrigger.shorthand.tsx
@@ -4,12 +4,8 @@ import { Button, Tooltip } from '@fluentui/react-northstar';
 const TooltipExampleDisabledTrigger = () => {
   return (
     <Tooltip
-      trigger={
-        <div style={{ display: 'inline-block' }}>
-          <Button disabled content="Hover me" />
-        </div>
-      }
-      content="This tooltip is set on the div wrapping the disabled button"
+      trigger={<Button disabledFocusable content="Hover me" />}
+      content="The button uses disabledFocusable property."
     />
   );
 };